fix: 메타 데이터 기본 조회 데이터 총 갯수 타입 수정

+++ src/main/java/visang/dataplatform/dataportal/model/dto/metadata/TableMetaInfoDto.java
@@ -23,5 +23,5 @@ public class TableMetaInfoDto {
     private String small_clsf_name;
 
     @Schema(description = "검색 결과 총 갯수")
-    private long total_num;
+    private int total_num;
 }
